[DEV-3483] moving checkbox to components
